requirements.txt
aiogram
fastapi
uvicorn
asyncpg
python-dotenv
pydantic-settings
apscheduler